Refactor schemas: update __init__.py to include ItemSearchCandidate and ItemSearchInput; enhance llm_search.py with confidence scoring and improve ItemSearchInput's to_prompt method for better JSON formatting
