+++ schemas/__init__.py
@@ -1,10 +1,12 @@
 """Pydantic schemas for the WMS application."""
 
-from .llm_search import ItemLocation
+from .llm_search import ItemLocation, ItemSearchCandidate, ItemSearchInput
 from .synthetic_data import BinGenerationOutput, ItemGenerationOutput
 
 __all__ = [
+    "BinGenerationOutput",
+    "ItemGenerationOutput",
     "ItemLocation",
-    "BinGenerationOutput", 
-    "ItemGenerationOutput"
+    "ItemSearchCandidate",
+    "ItemSearchInput"
 ]
